+++ controllers/spec/function.go
@@ -34,9 +34,8 @@ func MakeFunctionHPA(function *v1alpha1.Function) *autov2beta2.HorizontalPodAuto
 	objectMeta := MakeFunctionObjectMeta(function)
 	if function.Spec.Pod.HPAutoscaler == nil {
 		return MakeDefaultHPA(objectMeta, *function.Spec.Replicas, *function.Spec.MaxReplicas, function.Kind)
-	} else {
-		return MakeHPA(objectMeta, function.Spec.Pod.HPAutoscaler, function.Kind)
 	}
+	return MakeHPA(objectMeta, function.Spec.Pod.HPAutoscaler, function.Kind)
 }
 
 func MakeFunctionService(function *v1alpha1.Function) *corev1.Service {
